requirements.txt
streamlit
tensorflow==2.13.0
numpy==1.23.5
faiss-cpu
scikit-learn
